refactor(netlist): extract the lvnet renderer into render_lvnet.py
The full lvnet text renderer -- _TermLine, the handle system
(_LvnetHandles/_assign_lvnet_handles), every _render_lvnet_* function,
render_lvnet, and the lvnet-only helpers -- moves to a new render_lvnet.py
(1,433 lines). It imports only netlist_models / lvnet_grammar / graph helpers,
so it's a one-way leaf: nothing imports back into it except netlist.py's
re-export.

netlist.py re-exports render_lvnet and imports back the three helpers its
builder / old render_netlist still call (_quoted_frame_label,
_lvnet_const_value_str, _lvnet_literal_token); lvnet_parse and two test files
repoint their render-helper imports to render_lvnet. netlist.py 5,494 -> 4,122.

Byte-identical lvnet output; no import cycle; pyright src/ clean; the
netlist/lvnet test subset green (full suite via the hook).
